requirements.txt
Deprecated @ file:///work/perseverance-python-buildout/croot/deprecated_1698868068243/work
filelock @ file:///work/perseverance-python-buildout/croot/filelock_1701733993137/work
fsspec @ file:///croot/fsspec_1736274672181/work
gmpy2 @ file:///croot/gmpy2_1738085463648/work
importlib-metadata @ file:///croot/importlib_metadata-suite_1704813515092/work
Jinja2 @ file:///croot/jinja2_1737760107953/work
lightning-utilities==0.13.1
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///work/perseverance-python-buildout/croot/mpmath_1698864994882/work
networkx @ file:///croot/networkx_1737039604450/work
numpy @ file:///croot/numpy_and_numpy_base_1738078506782/work/dist/numpy-2.2.2-cp312-cp312-linux_x86_64.whl#sha256=189b4bed07855dc21202d8ad75c10aaa27068b030088b3cca054eb80a7925558
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
odak==0.2.6
opencv-python==4.11.0.86
opentelemetry-api @ file:///croot/opentelemetry-api_1724767675995/work
packaging==24.2
pillow==11.1.0
plyfile==1.1
setuptools==72.1.0
sympy==1.13.1
torch==2.6.0
torchmetrics==1.6.2
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
wheel==0.45.1
wrapt @ file:///croot/wrapt_1736540904746/work
zipp @ file:///croot/zipp_1732630741423/work
